Feat: Converted internal_reconcile_report_orphan to clean architecture.
Tests require changes.
