+++ lambda scripts/build.sh
@@ -94,6 +94,7 @@ let return_code=$?
 check_rc $return_code "ERROR: Failed to copy schema files to build directory."
 
 ## Create the zip archive
+echo "INFO: Creating zip archive ..."
 cd build
 zip -qr ../lambda_name.zip .
 let return_code=$?
