+++ tasks/copy_files_to_archive/bin/build.sh
@@ -116,6 +116,7 @@ let return_code=$?
 check_rc $return_code "ERROR: Failed to copy schema files to build directory."
 
 ## Create the zip archive
+echo "INFO: Creating zip archive ..."
 cd build
 zip -qr ../copy_files_to_archive.zip .
 let return_code=$?
