+++ tasks/extract_filepaths_for_granule/bin/build.sh
@@ -93,6 +93,7 @@ let return_code=$?
 
 check_rc $return_code "ERROR: Failed to copy schema files to build directory."
 ## Create the zip archive
+echo "INFO: Creating zip archive ..."
 cd build
 zip -qr ../extract_filepaths_for_granule.zip .
 let return_code=$?
