+++ tasks/post_copy_request_to_queue/bin/build.sh
@@ -109,6 +109,7 @@ let return_code=$?
 check_rc $return_code "ERROR: Failed to copy lambda files to build directory."
 
 ## Create the zip archive
+echo "INFO: Creating zip archive ..."
 cd build
 zip -qr ../post_copy_request_to_queue.zip .
 let return_code=$?
