+++ tasks/post_to_database/bin/build.sh
@@ -120,6 +120,7 @@ let return_code=$?
 check_rc $return_code "ERROR: Failed to copy schema files to build directory."
 
 ## Create the zip archive
+echo "INFO: Creating zip archive ..."
 cd build
 zip -qr ../post_to_database.zip .
 let return_code=$?
