+++ tasks/request_status_for_job/bin/build.sh
@@ -111,6 +111,7 @@ let return_code=$?
 check_rc $return_code "ERROR: Failed to copy lambda files to build directory."
 
 ## Create the zip archive
+echo "INFO: Creating zip archive ..."
 cd build
 zip -qr ../request_status_for_job.zip .
 let return_code=$?
